+++ src/statusBar.ts
@@ -77,7 +77,11 @@ const updateDiagnosticMessage = (
 	} else {
 		const diagnostic = diagnostics.find(({from, to}) => from <= main.head && to >= main.head) ?? diagnostics[0]!,
 			view = cm.view!;
-		msg.textContent = diagnostic.message;
+		if (diagnostic.renderMessage) {
+			msg.replaceChildren(diagnostic.renderMessage(view));
+		} else {
+			msg.textContent = diagnostic.message;
+		}
 		if (diagnostic.actions) {
 			msg.append(...(diagnostic.actions as ExtendedAction[]).map(({name, tooltip, apply}) => {
 				const button = elt('button', {type: 'button', class: actionSelector.slice(1)}, name);